refactor or electrode builders
- moved structure group builder to it's own file
- created base builder class  GroupedThermoDocsBuilder that just grabs group docs and thermo docs
- InsertionElectrodeBuilder is now a subclass
